requirements.txt
streamlit
pandas
matplotlib
langchain
langchain-google-genai
google-generativeai
python-dotenv
openpyxl
pytesseract
Pillow